requirements.txt
Django==1.7.4
argparse==1.2.1
cssselect==0.9.1
django-push-notifications==1.2.0
django-uuidfield==0.5.0
lxml==3.4.2
pyquery==1.2.9
requests==2.5.1
wsgiref==0.1.2
